Add files via upload
Business Scenario: 
A British company has been growing its business since its infancy stage. Their sales have been good as they deal with office infrastructure. However, the VP Sales is not very happy with the refunds that the company has been providing to its customers. The VP wanted a decision-driven method to be employed rather than a traditional approach to deriving a strategy. 

Objective: 
As a first step the VP planned to evaluate all the BI tools available in the market. Tableau was selected as it met his requirements. Using Tableau, he wanted his team members to give him the forecast of Sales for the next year as well as the probability of refunds category-wise for a year. This will help him plan his next steps to avoid the refunds. 
He wanted a dashboard that would contain the following: 
• A list of orders returned by the customers from the Top 10 Countries (in terms of refund) 
• Top 10 Countries mapped on the World Map that had most of the refunds 
• Predict the refund for next 1 year category-wise 

About the Data: 
The dataset being used is “global_superstore2016”. It consists of dimensions like Category, Sub-Category, OrderID, ProductID, OrderDate, etc.

Actions to be Performed to Create the Dashboard:
 • Create Hierarchies and Folders in the dataset provided
 • Generate a list of orders returned from customers and compare it to the original sales, sort the visualization in ascending order in terms of returned orders, for the top 10 countries in terms of refunds. 
• Map top 10 countries, with most of the refunds on the world map 
• Predict the returned sale with the lowest and actual forecast of the data 
• Add a URL action to represent details of the countries 
• Finally publish your work to Tableau Server/ Online